% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdParam.R
\name{BirdMorphParam}
\alias{BirdMorphParam}
\title{Morphological data from bird}
\usage{
BirdMorphParam(BMass, WSpan, WArea, C_db = 0.2, C_dpro = 0.2,
  ADensity = 1.23, grav = 9.8, k = 1.2)
}
\arguments{
\item{BMass}{Body mass of bird in kilograms (kg)}

\item{WSpan}{Wingspan (maximum distance between wingtips) of bird in meters (m)}

\item{WArea}{Wing area (area of fully stretched wings) of bird in square meters (\eqn{m^{2}}{m2})}

\item{C_db}{Body drag coefficient}

\item{C_dpro}{Profile drag coefficient}

\item{ADensity}{Air density}

\item{grav}{Acceleration due to gravity}

\item{k}{induced drag factor}
}
\value{
Morphological parameters of bird and other variables for power calculation.
}
\description{
This function creates a list of user-defined morphological description parameters of an
input bird. At least three parameters need to be specified: body mass, wing span and
wing area. This definition of the bird is then used by the other functions in the
package to estimate flight performance.
}
\details{
This function defines the morphological parameters of an input bird and some
other variables for power calculation. At least three bird parameters, body mass (\code{BMass}), wing span (\code{WSpan}) and
wing area (\code{WArea}),  need to be specified.

The default value for both body drag and profile drag coeficient: \code{C_db} and \code{C_dpro} is 0.2. A constant air density
of 1.23 \eqn{kg/m^{3}}{kg/m3} is used with acceleration due to gravity, \code{grav} of 9.8 \eqn{m/s^{2}}{m/s2}.
The induced drag factor, \code{k}, is also given as 1.2 as used in the \emph{Flight Program (Pennycuick, 2008)}.
All default arguments can be varied as suitable. If no arguments are defined, the default values will be used.
}
\references{
Pennycuick, C. J. (2008). \emph{Modelling the flying bird} (Vol. 5). Elsevier.
}
\seealso{
\code{FlightSpeedComponents}
}
