% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlightPerformance.R
\name{FlightPerformance}
\alias{FlightPerformance}
\title{Compute flight performance parameters}
\usage{
FlightPerformance(BirdParam, FlightSpeedComponents, TrueAirSpeed1,
  TrueAirSpeed2, C_l, C_t)
}
\arguments{
\item{BirdParam}{Basic morphological data of bird}

\item{FlightSpeedComponents}{Calculated flight speed components}

\item{TrueAirSpeed1}{True airspeed from function \code{TrueAirSpeed1}}

\item{TrueAirSpeed2}{Trueairspeed from function \code{TrueAirSpeed2}}

\item{C_l}{lift coefficent}

\item{C_t}{thrust coefficient}
}
\value{
Flight performance parameters made up of a dataset of all associated powers
which do not include infinite values. It also investigates the flight types as either straight steady, climbing or descending flight.
}
\description{
This function evaluates the power associated with the user-input flight trajectory
by using some other functions in this package. It also considers the 3D position of
the bird.
}
\details{
This function estimates all flight performance parameters by evaluating the various
types of aerodynamic forces acting during flight. It also checks for the feasibility of the
input trajectory and evaluates the amount of power required at each timestep. It also calculates the
mechanical and chemical power required for flight. The function makes use of the inputs for
\code{BirdMorphParam} and flight speed from \code{FlightSpeedComponents} as well as the 3D position
of the bird in flight. There is no default value for lift coefficient and thrust coefficient.
To accounT for wind effect, the true airspeed is used either from \code{TrueAirSpeed1} or
\code{TrueAirSpeed1}.
}
\references{
Norberg, U. M. (2012). Vertebrate flight: mechanics, physiology, morphology, ecology and
evolution (Vol. 27). Springer Science & Business Media.
}
\seealso{
\code{FlightSpeedComponents}
}
