% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlightSpeedComponents.R
\name{FlightSpeedComponents}
\alias{FlightSpeedComponents}
\title{Flight speed components}
\usage{
FlightSpeedComponents(t, x, y, z)
}
\arguments{
\item{t}{Timestep from data}

\item{x}{x-position from data}

\item{y}{y-position from data}

\item{z}{z-position from data}
}
\value{
Numeric value of flight Speed
}
\description{
The function first evaluates the x,y,z-components of the speed and then further evaluates the overall speed of
flight. This flight speed is the actual speed at which the bird moves in air relative to incident airflow.
}
\details{
\code{FlightSpeedComponents} sets up the x,y,z- components of the speed at which the bird flies
from pre-defined 3D position and timestep after which it estimates the actual speed at which the bird is flying.
}
\references{
...
}
