% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFlightTypes.R
\name{PlotFlightType}
\alias{PlotFlightType}
\title{Plot the input trajectory according to the types of flight}
\usage{
PlotFlightType(x, y, z, FlightPerformance)
}
\arguments{
\item{x}{x-position from data}

\item{y}{y-position from data}

\item{z}{z-position from data}

\item{FlightPerformance}{requires flighttype information from \code{FlightPerformance}}
}
\value{
3D scatterplot of flight types from the input trajectory
}
\description{
Plot the input trajectory according to the types of flight
}
\examples{
utils::data(trajectory)
trajectory = trajectory
bird = BirdMorphParam(BMass = 0.7710, WSpan = 0.98, WArea = 0.119, C_db = 0.1)
FSComponents=FlightSpeedComponents(t=trajectory[,4],
                   x=trajectory[,1], y=trajectory[,2], z=trajectory[,3])
Tairspeed = TrueAirSpeed1(FSComponents)
FPerformance = FlightPerformance(bird,FSComponents, Tairspeed,
                   C_l = 0.5, C_t = 0.1)

PlotFlightType(x=trajectory[,1], y=trajectory[,2], z=trajectory[,3], FPerformance)
}
