% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueAirspeed1.R
\name{TrueAirSpeed1}
\alias{TrueAirSpeed1}
\title{True airspeed (alternative 1)}
\usage{
TrueAirSpeed1(FlightSpeedComponents, WindSpeed.x, WindSpeed.y, WindSpeed.z)
}
\arguments{
\item{FlightSpeedComponents}{Components of speed of flight}

\item{WindSpeed.x}{x-compenent of speed of wind}

\item{WindSpeed.y}{y-compenent of speed of wind}

\item{WindSpeed.z}{z-compenent of speed of wind}
}
\value{
True air speed
}
\description{
This function is the first function for evaluating the effect of wind on flight speed.
It calculates true airspeed from user-defined x,y,z- windspeed components.
}
\details{
bbbb
}
\references{
bbb
}
