% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueAirspeed2.R
\name{TrueAirSpeed2}
\alias{TrueAirSpeed2}
\title{True airspeed (alternative 2)}
\usage{
TrueAirSpeed2(FlightSpeedComponents, V_WIND, theta, phi)
}
\arguments{
\item{FlightSpeedComponents}{Components of speed of flight}

\item{V_WIND}{Windspeed}

\item{theta}{Wind directions in degrees}

\item{phi}{Wind directions in degrees}
}
\value{
True airspeed
}
\description{
Calculate true airspeed from wind direction and speed
}
\examples{
x <- c(1:5)
y <- c(1:5)
z <- c(1:5)
t <- c(1:5)
FS.Components <- FlightSpeedComponents(t,x,y,z)
T.airspeed <- TrueAirSpeed1(FS.Components, 30, 15, 10)


}
