% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Powercurve.R
\name{PowerCurve}
\alias{PowerCurve}
\title{Plot mechanical Powercurve}
\usage{
PowerCurve(FlightPerformance, TrueAirSpeed1, TrueAirSpeed2)
}
\arguments{
\item{FlightPerformance}{Evaluated mechanical power}

\item{TrueAirSpeed1}{True airspeed}

\item{TrueAirSpeed2}{True airspeed}
}
\value{
Mechanical power curve
}
\description{
Plot mechanical Powercurve
}
\details{
This function plots mechanical power against flight speed.
}
