% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueAirspeed1.R
\name{TrueAirSpeed1}
\alias{TrueAirSpeed1}
\title{True airspeed (alternative 1)}
\usage{
TrueAirSpeed1(FlightSpeedComponents, WindSpeed.x, WindSpeed.y, WindSpeed.z)
}
\arguments{
\item{FlightSpeedComponents}{Components of speed of flight}

\item{WindSpeed.x}{x-compenent of speed of wind}

\item{WindSpeed.y}{y-compenent of speed of wind}

\item{WindSpeed.z}{z-compenent of speed of wind}
}
\value{
True airspeed
}
\description{
This function is the first function for evaluating the effect of wind on flight speed.
It calculates true airspeed from user-defined x,y,z- windspeed components.
}
\details{
If x,y,z-components of wind speed are not given, flight speed obtained from \code{FlightSpeedComponents}
is assumed as true airspeed.
}
\examples{
x <- c(1:5)
y <- c(1:5)
z <- c(1:5)
t <- c(1:5)
wx <- seq(0.05,1,length.out=5)
wy <- seq(0.06,1.2,length.out=5)
wz <- seq(0.07,1.5,length.out=5)
FS.Components <- FlightSpeedComponents(t,x,y,z)

# When wind speed components are not given
T.airspeed <- TrueAirSpeed1(FS.Components)

# When wind speed components are given
T.airspeed <- TrueAirSpeed1(FS.Components, wx, wy, wz)

}
