% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueAirspeed2.R
\name{TrueAirSpeed2}
\alias{TrueAirSpeed2}
\title{True airspeed (alternative 2)}
\usage{
TrueAirSpeed2(FlightSpeedComponents, V_WIND, theta, phi)
}
\arguments{
\item{FlightSpeedComponents}{Components of speed of flight}

\item{V_WIND}{Windspeed}

\item{theta}{Wind directions}

\item{phi}{Wind directions}
}
\value{
TrueAirSpeed
}
\description{
Calculate true airspeed from wind direction and speed
}
\details{
bbb
}
\references{
bbb
}
